change from int to float

+++ Encoder/Generator.py
@@ -1,7 +1,6 @@
 import torch
 
 def generator(nr_codewords, device):
-    bits = torch.randint(0, 2, size=(nr_codewords, 1, 4), dtype=torch.int, device=device)
-    # bits = torch.tensor([[[1, 1, 0, 1]], [[0, 1, 0, 0,]]], dtype=torch.int, device=device)
+    bits = torch.randint(0, 2, size=(nr_codewords, 1, 4), dtype=torch.float, device=device)
 
     return bits
